# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Faarooq Backend API",
    "version": "0.1.0",
    "description": "OpenAPI spec for the Twitter-like backend (Express + Supabase)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://faarooq-backend.onrender.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "responses": {
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until reset" },
          "X-RateLimit-Limit": { "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
          "X-RateLimit-Reset": {
            "schema": { "type": "integer" },
            "description": "Unix epoch seconds when window resets"
          }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            },
            "required": ["code"]
          }
        },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "handle": { "type": "string" },
          "display_name": { "type": "string" },
          "bio": { "type": "string" }
        },
        "required": ["id", "handle"]
      },
...
    }
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "summary": "Get current user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "put": {
        "summary": "Update current user profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "display_name": { "type": "string" }, "bio": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "409": {
            "description": "Handle already taken",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/api/users/me/avatar": {
      "post": {
        "summary": "Attach/update avatar (after uploading via media upload-url)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "avatar_url": { "type": "string", "format": "uri" } },
                "required": ["avatar_url"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/posts": {
      "get": {
        "summary": "Get timeline feed (posts + repost activities; excludes soft-deleted posts)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK" },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "post": {
        "summary": "Create post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "text": { "type": "string", "maxLength": 280 } },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/reposts": {
      "post": {
        "summary": "Repost a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      },
      "delete": {
        "summary": "Unrepost a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/follows": {
      "post": {
        "summary": "Follow a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "followee_id": { "type": "string", "format": "uuid" } },
                "required": ["followee_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "delete": {
        "summary": "Unfollow a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "followee_id": { "type": "string", "format": "uuid" } },
                "required": ["followee_id"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/likes": {
      "post": {
        "summary": "Like a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "delete": {
        "summary": "Unlike a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/media": {
      "post": {
        "summary": "Attach uploaded media to a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "post_id": { "type": "string", "format": "uuid" },
                  "path": { "type": "string" },
                  "media_type": { "type": "string", "enum": ["image", "video", "gif"] }
                },
                "required": ["post_id", "path"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "summary": "List notifications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK" },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  }
}